--- a/package.json
+++ b/package.json
@@ -25,7 +25,7 @@
     "build:publish-types": "tsc -p tsconfig.publish.types.json && node ./buildTypes.js",
     "build:publish-cjs": "tsc -p tsconfig.publish.json",
     "build:publish-esm": "tsc -p tsconfig.publish.json --module es2015 --moduleResolution node --outDir lib/esm",
-    "webbuild": "browserify lib/fast-check.js --s fastcheck | uglifyjs -cm > lib/bundle.js",
+    "webbuild": "browserify lib/fast-check.js --s fastcheck | terser -c -m > lib/bundle.js",
     "build:prod": "yarn prebuild && yarn build:publish-cjs && yarn build:publish-esm && yarn build:publish-types && yarn webbuild && node postbuild/main.cjs",
     "watch": "tsc -w",
     "test": "jest --config jest.unit.config.cjs --coverage",
@@ -52,11 +52,10 @@
   },
   "homepage": "https://github.com/dubzzz/fast-check#readme",
   "engines": {
-    "node": ">=0.12"
+    "node": ">=8.0.0"
   },
   "dependencies": {
-    "pure-rand": "^2.0.0",
-    "tslib": "^2.0.0"
+    "pure-rand": "^2.0.0"
   },
   "devDependencies": {
     "@microsoft/api-extractor": "^7.8.10",
@@ -79,10 +78,10 @@
     "rollup-plugin-commonjs": "^10.1.0",
     "rollup-plugin-node-resolve": "^5.2.0",
     "source-map-support": "^0.5.19",
+    "terser ": "^4.4.3",
     "ts-jest": "^25.5.1",
     "tslint-config-prettier": "^1.18.0",
-    "typescript": "^3.9.5",
-    "uglifyjs": "^2.4.11"
+    "typescript": "^3.9.5"
   },
   "keywords": [
     "property-based testing",