--- a/src/check/arbitrary/UuidArbitrary.ts
+++ b/src/check/arbitrary/UuidArbitrary.ts
@@ -1,10 +1,9 @@
-import { StringPadStart } from '../../utils/polyfills';
 import { Arbitrary } from './definition/Arbitrary';
 import { integer, nat } from './IntegerArbitrary';
 import { tuple } from './TupleArbitrary';
 
 /** @hidden */
-const padEight = (arb: Arbitrary<number>) => arb.map((n) => StringPadStart(n.toString(16), 8, '0'));
+const padEight = (arb: Arbitrary<number>) => arb.map((n) => n.toString(16).padStart(8, '0'));
 
 /**
  * For UUID from v1 to v5