… src/utils/polyfills.ts
… test/legacy/main.js
… test/unit/utils/polyfills.spec.ts
… yarn.lock