**Merge of PR #515 to master**

---

## Why is this PR for?

Remove support for old versions of Node.

Removing the support brings a clear uplift in terms of performances, here are the results we measured on [RunKit](https://runkit.com/dubzzz/chrono-fast-check) by running the same code against `fast-check@1.26.0` and `fast-check@2.0.0-alpha.1.26.0.0`:
- synchronous property: from 0.30ms to 0.26ms (-14% on median time)
- asynchronous property: from 0.54ms to 0.36ms (-34% on median time)

```js
// synchronous property
(fc, seed) => {
  const start = performance.now();
  fc.assert(fc.property(fc.nat(), fc.nat(), (a, b) => true), {seed});
  return performance.now() - start;
}
// asynchrounous property
async (fc, seed) => {
  const start = performance.now();
  await fc.assert(fc.asyncProperty(fc.nat(), fc.nat(), async (a, b) => true), {seed});
  return performance.now() - start;
}
```

Most of the uplift might be related to the fact that neither `yield` nor `async`/`await` need to be transpiled if we drop support for older releases on nodes.

## In a nutshell

❌ New feature
❌ Fix an issue
❌ Documentation improvement
✔️ Other: *remove support*

(✔️: yes, ❌: no)

## Potential impacts

Code will no longer run on old versions of node.